Calculate heritability

Calculate the heritability based on the fitted model. The heritability is calculated as described by Atlin et al. E.g. for a model with trials nested within locations, which has a random part that looks like this: genotype + genotype:location + genotype:location:trial the heritability is computed as

$$\sigma_G^2 / (\sigma_G^2 + \sigma_L^2 / l + \sigma_{LT}^2 / lt + \sigma_E^2 / ltr)$$ In this formula the \(\sigma\) terms stand for the standard deviations of the respective model terms, and the lower case letters for the number of levels for the respective model terms. So \(\sigma_L\) is the standard deviation for the location term in the model and \(l\) is the number of locations. \(\sigma_E\) corresponds to the residual standard deviation and \(r\) to the number of replicates.

herit(varComp)

Arguments

varComp
An object of class varComp.

References

Atlin, G. N., Baker, R. J., McRae, K. B., & Lu, X. (2000). Selection response in subdivided target regions. Crop Science, 40(1), 7–13. https://doi.org/10.2135/cropsci2000.4017

See also

Examples

## Fit a mixed model.
geVarComp <- gxeVarComp(TD = TDMaize, trait = "yld")

## Compute heritability.
herit(geVarComp)

#> [1] 0.3489369