statgenGxE: Genotype by Environment (GxE) Analysis

Analysis of multi environment data of plant breeding experiments following the analyses described in Malosetti, Ribaut, and van Eeuwijk (2013), doi:10.3389/fphys.2013.00044 . One of a series of statistical genetic packages for streamlining the analysis of typical plant breeding experiments developed by Biometris. Some functions have been created to be used in conjunction with the R package 'asreml' for the 'ASReml' software, which can be obtained upon purchase from 'VSN' international (https://vsni.co.uk/software/asreml-r).

See also

Author

Maintainer: Bart-Jan van Rossum bart-jan.vanrossum@wur.nl

Other contributors:

Fred van Eeuwijk (ORCID) [contributor]

Martin Boer [contributor]

Marcos Malosetti (ORCID) [contributor]

Daniela Bustos-Korts (ORCID) [contributor]

Emilie J Millet (ORCID) [contributor]

Joao Paulo (ORCID) [contributor]

Maikel Verouden (ORCID) [contributor]

Willem Kruijer (ORCID) [contributor]

Ron Wehrens (ORCID) [contributor]

Choazhi Zheng (ORCID) [contributor]